+++ backend/src/main/java/com/betacom/backend/request/OrdineReq.java
@@ -7,7 +7,7 @@ public class OrdineReq {
     private Integer id;
     private String data;
     private String stato;
-    private Integer prezzoTotale;
+    private Double prezzoTotale;
     private Integer qty;
     private Integer idUtente;
     private List<ProdottiOrdiniReq> prodOrdiniReq;
@@ -15,7 +15,7 @@ public class OrdineReq {
     public OrdineReq() {
     }
 
-    public OrdineReq(Integer id, String data, String stato, Integer prezzoTotale, Integer qty,
+    public OrdineReq(Integer id, String data, String stato, Double prezzoTotale, Integer qty,
         Integer idUtente, List<ProdottiOrdiniReq> prodOrdiniReq) {
         this.id = id;
         this.data = data;
@@ -50,11 +50,11 @@ public class OrdineReq {
         this.stato = stato;
     }
 
-    public Integer getPrezzoTotale() {
+    public Double getPrezzoTotale() {
         return prezzoTotale;
     }
 
-    public void setPrezzoTotale(Integer prezzoTotale) {
+    public void setPrezzoTotale(Double prezzoTotale) {
         this.prezzoTotale = prezzoTotale;
     }
